refactor: updates logic for editable BLIs
