+++ frontend/src/hooks/budget-line.hooks.js
@@ -1,5 +1,4 @@
 import { useSelector } from "react-redux";
-import { BLI_STATUS } from "../helpers/budgetLines.helpers";
 
 /**
  * @typedef {Object} BudgetLine
@@ -8,24 +7,6 @@ import { BLI_STATUS } from "../helpers/budgetLines.helpers";
  * @property {number} created_by - The ID of the user who created the budget line.
  * @property {boolean} in_review - Whether the budget line is in review.
  */
-
-/**
- * This hook returns true if the budget line is in a status that is editable.
- * @param {BudgetLine} budgetLine - The budget line object.
- * @returns {boolean} - Whether the budget line is in a status that is editable.
- * @example
- * const isBudgetLineInEditableStatus = useIsBudgetLineEditableByStatus(budgetLine);
- */
-export const useIsBudgetLineEditableByStatus = (budgetLine) => {
-    const isBudgetLineDraft = budgetLine?.status === BLI_STATUS.DRAFT;
-    const isBudgetLinePlanned = budgetLine?.status === BLI_STATUS.PLANNED;
-    const isBudgetLineExecuting = budgetLine?.status === BLI_STATUS.EXECUTING;
-    const isBudgetLineInReview = budgetLine?.in_review;
-    const isBudgetLineInEditableStatus =
-        (isBudgetLineDraft || isBudgetLinePlanned || isBudgetLineExecuting) && !isBudgetLineInReview;
-
-    return isBudgetLineInEditableStatus;
-};
 
 /**
  * This hook returns true if the logged in user is the creator of the budget line.
