+++ frontend/src/pages/budgetLines/list/BudgetLinesItems.hooks.js
@@ -1,6 +1,6 @@
 import React from "react";
-import { useSearchParams } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { useSearchParams } from "react-router-dom";
 
 export const useBudgetLinesList = () => {
     const activeUser = useSelector((state) => state?.auth?.activeUser);
